requirements.txt
certifi==2019.6.16
chardet==3.0.4
dj-database-url==0.5.0
Django==2.2.2
django-debug-toolbar==1.11
django-taggit==0.23.0
gunicorn==19.9.0
idna==2.8
oauthlib==3.0.2
Pillow==6.1.0
pkg-resources==0.0.0
psycopg2-binary==2.8.3
py-trello==0.15.0
python-dateutil==2.8.0
pytz==2019.1
requests==2.22.0
requests-oauthlib==1.2.0
sentry-sdk==0.10.2
six==1.12.0
sqlparse==0.3.0
Unidecode==1.0.23
urllib3==1.25.3
whitenoise==4.1.2
